+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
  - Sample Editor: Dither preference.
 
+### Changed
+ - Sample Editor: Moved dither preference to main window.
+
 ## [1.11.5] - 2021-06-16
 ### Changed
  - Sample Editor: Tweaked resampler lowpass-filter settings to preserve more treble.