+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+ - Sample Editor: Dither preference.
+
+### Changed
+ - Sample Editor: Moved dither preference to main window.
 
 ## [1.11.5] - 2021-06-16
 ### Changed