+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [1.11.6] - 2021-08-18
 ### Fixed
  - Sample Editor: Dither preference.
  - Sample Editor: Allow filenames less than 3 characters in length.